// FSLogistic.Web/ClientApp/src/core/models/update-customer.model.ts
export class UpdateCustomer {
    id:number;
    code:string;
    name: string;
    q1: number;
    q2: number;
    q3: number;
    q4: number;
    q5: number;
    q6: number;
    q7: number;
    q8: number;
    q9: number;
    q10: number;
    q11: number;
    q12: number;
    qtd: number;
    qgv: number;
    qbth: number;
    qtb: number;
    qtp: number;
    qbtn: number;
    qpn: number;
    qcc: number;
    qhm: number;
    qbc: number;
    qnb: number;
    qcg: number;
}